doctor/backend/extraction-service/requirements.txt
# Medical Report Extraction Service Dependencies

# FastAPI framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# PDF Processing
pymupdf>=1.23.0
pdfplumber>=0.10.0

# OCR
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data validation
pydantic>=2.0.0

# HTTP client (for testing)
httpx>=0.24.0

# LLM Integration
groq>=0.4.0
python-dotenv>=1.0.0

# Google Gemini API
google-generativeai>=0.3.0

# CSV/EHR Processing
pandas>=2.0.0

# For Windows: Install Tesseract OCR separately
# Download from: https://github.com/UB-Mannheim/tesseract/wiki
# Add to PATH or set pytesseract.pytesseract.tesseract_cmd
